Remove .sty file before building it
If the .sty file already exists when the installer file is processed,
the installer will ask the user for confirmation before overwriting it.
That’s inconvenient for an automatic install, so let’s simply remove the
file beforehand.

I had actually added this rule once already while working on the
Makefile, but then couldn’t remember the reason for it anymore and
removed it before committing the first version of the Makefile. So if
we’re adding it back now, let’s document the reason :)