Remove misleading line breaks
I added some line breaks to the StackOverflow answer where I got the
helper macro from, but had misinterpreted this part. The \relax is not
the “then” part of the \if, it’s the second part of the condition: if #1
is equal to \relax, do nothing (empty “then” part), else do the rest of
the macro.